global_requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
pillow==12.2.0
pycparser==2.21
PyJWT==2.12.1
PyMySQL==1.1.2
sqlparse==0.5.5
tzdata==2026.1
